Bug#36715372 tls session reuse does not work with read-write splitting
Connections in the connection pool get closed without waiting for the
server to close the connection which affects the TLS session-reuse
cache.

When a TLS session shall be reused the session must be closed by waiting
for client and the server to announce the close of their TLS session by
calling SSL_shutdown() twice.

Once for the client and a 2nd time when the server returned the close or
the socket was closed.

This affects connection sharing and read-write splitting as both place
let the connection-pool close their connections.

Change
======

- refactored ConnectionPool::add() on top of
  ConnectionPool::add_if_not_full() as the share the most of the code.
- added a proper async socket shutdown for pooled connections.

Change-Id: Ic8849a576e2d77279e7e789c033f0d21642308c6