Bug#36703827 Split rpdhpgtest into smaller tests
The unit test rpdhpgtest currently takes ~15 to complete.
This makes it time out in pushbuild.
Prior to this patch
    Bug#36591084 Remove usage of gtest_discover_tests
it was executed as 826 separate tests.

This patch splits rpdhpgtest into rpdhpgtest_1 rpdhpgtest_2 ....
one for each source file.

We also adjust the timeout for the 'test-unit' cmake target to about
double the time taken by the most time consuming of the rpdhpgtest_n
executables.

Change-Id: Iec40904192892c788cbf40ed7e5647dfebc8a35d