Bug#35200385 The parallel tablespace scan in 8.0 doesn't work as expected
Problem:
Validate_files::validate opens ibd files in multiple threads, but
this file operation is protected using a mutex, which is originally
used for protecting access to in-memory data structures during
validation. This is unnecessary, as each thread can freely open a
separate ibd file.

Solution:
Narrow down the scope protected with said mutex, making sure that
the ibd files are checked outside that scope
The mutex now only used when accessing the vector of moved
tablespaces.

Change-Id: I094ec539b230efa07c7c2757407160e698e38c0c