Bug#36701396 Sessions reported to still be connected at end of tests
MTR's check of test cases is using I_S.processlist, and since the
patch for Bug#34930219, cached values are used to determine if
connections are alive. Sometimes, sessions being disconnected can
be delayed in a way that is causing the test case check to report
connections still being alive.

The fix is to update the cached alive value during disconnect in
THD::release_resources().

Change-Id: Ic7c310e816df6c8f6d65c65dbc0ae624ec1485d0