Bug#36705030: prepared statement encounter performance degradation compared with normal select
When using a dynamic parameter the optimizer was not able to change
a ref access to a range access. The reason for this was a returned
dependency map from calc_length_and_keyparts() that distinguished
true constant values and dynamic parameters.

The fix is to ignore used_tables bit for dynamic parameters.

Change-Id: Idc7acf2a6585d92f7cd7b2b49c4054e2db2fb2c7