Bug#36705061: Case query assertion error
The query asserts because we have a generated column description
with a temporal type and a fractional seconds precision that is 31,
which is an invalid precision value. The invalid precision comes from
type aggregation, where one of the underlying items to the CASE
expression is a NULL expression, which has "decimal" precision = 31,
meaning an undefined precision.

The problem is fixed by not aggregating types from expressions that
have a NULL type. Notice that numeric and string arguments can never
be directly aggregated into a temporal type, so they need not be
handled.

Change-Id: I4c677dcf0e0e3024631710bc4807f28d24a61e8e