Bug#35767731 Replace all use of 'utf8' and '_utf8' [std_data noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' in std_data files.

The patch also removes the unused std_data/bug18311187_account.TRG

The patch also unzips two data files, and fixes the use of 'utf8'.
These data files are left uncompressed, which increases readability,
and simplifies the tests using them.

Change-Id: Ifa3dc328f942cb4a86e89bb22403edb49079d15a