Bug#36705312 recent regression: heap corruption: Assertion failed: dptr <= buffer->ptr() + buffer->alloced_length() - part 1 of 2 [noclose]
Also fixes duplicate Bug#36709784 mysqld crash at pack_rows::StoreFromTableBuffers

This is a regression from
Bug#36075756 ASAN crash on MaterializeIterator<Profiler>::load_HF_row_into_hash_map() [2/2]

valgrind gives an error when running
- sysschema.pr_statement_performance_analyzer
- the repro in this issue

This line in CalculateColumnStorageSize makes a jump decision
unitialized data:
    return field_blob->is_null() ? 0 : ....

Testing blobs for is_null in CalculateColumnStorageSize is only safe
if we have just read a row (or otherwise put a valid value in there)
so the Field buffer is in a valid state. This was not always the case
for other uses of ComputeRowSizeUpperBound than our own (as evidenced
by a failed valgrind run of
sysschema.pr_statement_performance_analyzer), which originated from a
call in AggregateIterator::AggregateIterator. So, do not check nulls
unless we know we have a valid value, signalled by the new argument to
ComputeRowSizeUpperBound.

In a follow-up patch, we will try to tease apart the two usages of
this method:
     - a worst case size for all rows of a given schema
     - the size of an actual row.

for fixed size rows, the two are equal, but for blobs we need to
differentiate. The usage in for example
AggregateIterator::AggregateIterator reads garbage data trying to
compute the actual size for blobs with no actual row in the record
buffers. This apparently hasn't caused any issues till now. but with

Bug#36075756 ASAN crash on
             MaterializeIterator<Profiler>::load_HF_row_into_hash_map()
             [2/2]

we started testing values (in an if statement) of such garbage data,
and valgrind complained.

Change-Id: I0866fad692b59c595c3faef48db6f490ef9f8593