Bug#36458177: Mysqld failure THD::rollback_item_tree_changes()
The problem is a stack-use-after-return for a recorded item address,
detected when item changes are rolled back after query optimization
and execution. Notice that the problem is suppressed during MTR runs
of an ASAN build when the --sanitize option is used.

The problem occurs in function remove_eq_conds() when calling
remove_eq_conds() recursively. The replaced item is returned in local
variable new_item. However, remove_eq_conds() may call fold_condition()
which may record its changes using THD::change_item_tree(), which will
then record the address of new_item, which is obviously invalid after
returning from the function. remove_eq_conds() does not need to record
addresses because it operates on a copy of the condition structure
(the AND/OR-tree), which need no rollback after optimization.

The solution to the problem is to replace the item pointer directly in
the list object that contains it, instead of using a local variable and
calling li.replace() explicitly.

Change-Id: I223acda5df955317e432ad4596b031f91bfea723