Bug#36652127 InnoDB: FTS table optimization not working correctly with collation
Symptom:
FTS index optimization sometimes works incorrectly with tokens that
are equal in collation order.

Cause:
When creating a list of words to optimize, tokens that are equal in
collation order, but byte-wise different, are treated as separate words,
except when they differ only by case. However, when a word from that
list is used to select index entries, all entries containing tokens
equal by collate order are selected. This means that they can be
processed by optimization more than once, in some cases leading to
incorrect behavior.

Fix:
When creating the list of words to be optimized, treat words that are
equal in collation order as duplicates. Since the words are returned in
collation order (due to ORDER BY clause), they are guarantted to be
adjacent, meaning only the first of the collation-equal versions will be
added to the list.

Change-Id: I8d012f997a7d41b5f7fb5213071d77e3168a8852