Bug#35877063: Subquery refcount assertion error
The problem occurs when removing a scalar subquery with a HAVING clause
after having decided that the subquery is part of a predicate that can
be eliminated. The HAVING clause points to an aggregation object
(Item_sum_min), indirectly accessed through an Item_aggregate_ref
that was added by Item::split_sum_func2().
When removing the subquery, the Item_aggregate_ref's ref count is
decremented, however it has never been incremented, hence the assert.

The fix to this poblem is to increment the object's reference count
when it is created.

Change-Id: I19121bd5b9f98071384053668586991b2ccd1dd6