BUG: Revaled by metadata_cluster_*.test that use GR MultiMaster mode.
Description
===========
At test cleanup, router that is configured to monitor its
usage-statistics, tries to write them to metadata-schema.

There are concurent DML and DDL queries executed by cleanup
and statistic-update, this breaks GR cluster.

Fix
===
Test is going to disable MRS plugin fiest and the do the clienup.
This should be done only when GR is installed and it works in MM
mode.

Change-Id: I9da2671753b1e51c8a87daffb6f66e4450ce0eca