Bug#36734676 removed unused code around QuitSender
In the past the QuitSender has been used to terminate the connection
between Router and Server in a controlled way when connection-sharing
was used.

But with "connection-sharing without connection-pool",
connections are stashed whenever the connection becomes sharable and
don't need to be closed anymore if the pool is full.

That removes the need for the QuitSender and the "pool_backend" stages.

Change
======

- remove the pool_backend, switch_backend stages of
  StmtPrepareForwarder.
- remove the include of quit_sender.h
- added test for more TLS resumption around rw-splitting and the
  StmtPrepare "force-switch-to-primary" scenario.

Change-Id: Id3e11c19300860361df0eb31ea18c9bc839f1561