Allowing configuring the MRS so that when starting it waits for a configured number of seconds for the service_metadata and service_data users to be granted expected access on the MySQLServer.
[mysql_rest_service]
wait_for_metadata_schema_access=10

will wait for a maximum of 10 seconds before erroring out
if the configured users do not have expected GRANTs.

This is usefull for mtr tests where the server and router
are launched in parallel and only the the tests are createing MRS
schema and grant proper access.

Change-Id: I025a4adf703a2d3b41ec62bff18d5679258e6a22