Bug#36514339 ROLLUP misses a summary NULL
Even with an empty result set, ROLLUP should give a single grouping
row with NULLs for all group by expressions and grand total for
aggregates, e.g.  0 for COUNT(*).

We have verified that this is in accordance with the SQL standard.

This is what PostgreSQL and DuckDB does as well, whereas SQL server
and OracleDB (19c) server return an empty result set.

Change-Id: Ib86a6b725e08a6a70efe2f8f4927e8041b6a0d80