BUG: Fix the MRS not resuming handling the requests after the schema has been removed and recreated
The new MTR approach is to reuse the existing MTR process between the tests
if the congiuration does not change.
That exposed a problem with MRS. Each test creates a MRS schema and
some objects for testing and removes them when finished. Turns out
that MRS does not notice the schema being recreated and stops
handling any requests between the tests.

This patch fixes that by resetting the MRS state once the error
(no schema or table does not exist) is discovered.

Change-Id: I515f6b7cf0945a8e31f7a4bce5ecda9fc7deea26