Bug#36730830  Assertion failure: log0files_governor.cc:1868:it != log.m_files.end() thread 140
Description:
------------
  When redo log capacity reduces, the redo log files are truncated using
log_files_truncate. This function calculates the new size, does the
truncate and updates the file handles. When updating the file handle, we
find that the write_lsn is not contained in the redo log files.

Background:
-----------
  The Redo log module uses convention similar to std library: the ranges
are conceptually open on the right end. For example, a file contains
[file.start_lsn, file.end_lsn) fragment of the redo log and the written
portion of the redo log is [LOG_START_LSN, log_sys->write_lsn).
  When truncating, we ensure not to remove already written data by
making sure that the new file.end_lsn is at least equal to
log_sys->write_lsn. We take other factors into account, too, like
aligning up to a block boundary or not making the file smaller than
requested, so most of the time we truncate somewhere above
log_sys->write_lsn.

Issue:
------
  During truncation, in the rare case where the log_sys->write_lsn is a
good enough number, the file is truncated such that file.end_lsn is
equal to log_sys->write_lsn. So, when reopening the file after
truncation, none of the files contains the log_sys->write_lsn.
Conceptually, the log_sys->write_lsn belongs to a future file which
does not exist yet.
  The situation where log_sys->write_lsn is equal to file.end_lsn
normally occurs when the last write finished at the boundary. In these
cases, the handle is already open and we request the creation of a new
file. Truncation, by contast, tries to find the file to open but can't
find any.

Fix:
----
  When calculating the new_size of the current redo log file, we add 1
to handle this rare case. This will ensure that write_lsn is properly
contained within the file after truncation. In particular, if write_lsn
was equal to file.end_lsn, no truncation will take place.

Change-Id: I908fe32ffc88b8d89d6a2ad3f236ad3c88d73cbd