Bug#36636836: mysql_server option is not registered upon reinstalling option_tracker component
The weak reference class was not holding active service references to
the underlying service after calling the init() method.
And, once the init function for the weak reference is called there was
a flag that prevented further calls.
This makes it is possible to unload and load the component implementing the
weak reference and its state will be reset but the caller of the weak
reference will not notice it.
In this case, at deinit time (since the state is reset), deinit() functor
can fail and this put the server in an irrecoverable state that can only
be fixed by killing it.

The fix is to implement 2 modes for weak_reference:

1. hold the reference mode (default):
In this mode the weak_reference class will store the service reference
to the weak function implementation on first use and will then use this
consistently with the deinit functor and release it in the deinit()
for the weak reference class. This of course means that the component
implementing the service that weak_reference is used on (in this case
the option tracker component) cannot be unloaded until the component
using the weak reference is unloaded. But that's the right thing to do
since the state deinitialization must be performed in the right
sequence.

2. release the reference, but keep the service loaded callback firing.
In this mode the weak_reference class will acquire and release reference
to the weak service implementation as before. But it will not prevent
the service loaded callback from calling the init functor multiple times.
This obviously has the disadvantage that init functor might be called
several times if the component using a weak reference is loaded and
 unloaded. This makes this mode not really useful for run-of-the-mill
components. But it's perfect for the server component registering itself
since one would want to allow the weak service implemetation component
(option_tracker in this case) to be unloaded and loaded during the
 lifetime of the server if nothing else prevents it.
And there's no possibility to call the weak_reference init functor
multiple times in this case, since this is governed by the server
component and can only happen if the option tracker is reloaded.

Added a boolean for the 2 modes to weak_reference::init()
Made the server use mode 2 by passing false.
Made the default value for weak_reference::init parameter such that
mode 1 is a default. Hence all current uses of weak_reference will
switch to mode 1.

Added a test case for mode 1 and mode 2.
Fixed doxygen build
Addressed Marc's comments

Change-Id: Ide0fb709ac1bf9927601bcd8a4bc47ddea78e378