Bug#36750146 Ndb : Graceful node shutdown does not avoid event duplicates
Graceful node stop is intended to carefully handover
responsibility for forwarding event streams from the stopping
node to still-running nodes, so that there are no duplicate
events sent to event subscribers as a result of an unexpected
node failure.

However, this was not working correctly which could lead
to issues with event merge and other event consumers even
when performing a graceful node stop.

The problem is fixed and two new tests covering the problem
with Graceful stop are added :

  test_event -n EventConsumer_Graceful
  test_event -n MergeEventConsumer_Graceful

Change-Id: I2effa2e0908c920e48ac235a5e0f3a1fce10413a