Bug#36444257: `false' in operator at sql/sql_executor.cc
A const item that is part of group by and not found in select list is not
added as a hidden item to the field's list post the fix for Bug#34951115.
This is not taken into consideration while checking for replacements for
expressions involving ROLLUP operation during temp table creation. So
server fails a check.
Fix is to check if the item that is not found in the field's list
is a const item. If so, the same item will be used as a replacement.

Change-Id: Ifca0618ef42930018def7a787f8c9aa3cd7d8be8