Bug#35767731 Replace all use of 'utf8' and '_utf8' [rpl noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' for the rpl test suite.

Some .result files from other test suites have also been updated.
This is caused by inclusion of common .inc and/or .test files.

Change-Id: I635f5276e18c6b3d3b267a2648559a00cdbb4691