Bug#36705312 recent regression: heap corruption: Assertion failed: dptr <= buffer->ptr() + buffer->alloced_length() - part 2 of 2
Refactoring of ComputeRowSizeUpperBound and its usage.  In summary of
analysis of ComputeRowSizeUpperBound (see gerrit), we never need the
worst case size for blobs (~4Gb), we either want:

    1) computed row size, excluding any blobs. No actual data need be
    present in the tables.

    2a) actual size, including blobs, heed any NULL flag. Requires
    valid data.

    2b) actual size, including blobs, but ignore any NULL flag.
    Requires valid data. This is a special case for NULL complemented
    rows.  Requires passing down "tbl.store_contents_of_null_rows &&
    table->has_row()"

The latter two are encapsulated inside the implementation of
ComputeRowSizeUpperBound.

So, we have split ComputeRowSizeUpperBound into two:

 - ComputeRowSizeUpperBoundSansBlobs - does not require valid data in
                                       record buffers.  Calling this
                                       is the record buffer has blobs
                                       will assert in debug mode

 - ComputeRowSizeUpperBound - requires valid data in record buffer
                              because it looks at blobs as
                              well. It may or may not heed the NULL
                              flag.

CalculateColumnStorageSize has a new parameter: skip_blob_null_check.

Note we tried to assert TABLE::has_row for the case where we (know we)
have valid data in the table buffers, but that failed, since a failed
read sets ::has_row() to false and we need to save record buffers in
such a case, cf. AggregateIterator::read(). So, we have to reply on
user (of ComputeRowSizeUpperBound) discretion to get this right.

Change-Id: I8fd6752d58fcebbda1f6ab5d7df6c0307229c4df