Bug#36368181: Assertion failure in operator() at               sql/sql_derived.cc
If an (underlying) expression involved in condition
pushdown had one view reference which is also a outer
reference and one local reference then the consistency
checks for "depended_from" information failed while
cloning the condition.
E.g If an expression that is being cloned is of the
form "dt1.f1 + t1.f1" where "dt1.f1" is a view reference
that is also an outer reference then currently while
collecting the "depended_from" information for the
field, we first save the "depended_from" and "context"
information of the "Item_ref" object (dt1.f1) while
traversing and use it for the underlying field(f1).
However, we do not reset the "depended_from" information
which leads to using the same information for the next
column "t1.f1" which is a local reference. Because of
this inconsistency, an assert fails.

Solution is to traverse the item tree both prefix and
postfix. While traversing prefix we set the "Item_ref"
object to be used for setting the "depended_from" and
"context" information for the underlying columns of
the ref object and while traversing postfix, we clear
the "Item_ref" object to be used.

Change-Id: I2961df5933efe9f1cc275fb7f80e94e734170ec1