Bug#36739383 Assertion `false' materialize_iterator::SpillState::read_next_row_secondary_over - part 1/4 [noclose]
With production builds, the server hangs.

The problem here is

    a) we encounter space overflow in the in-memory hash table during
       a set operation (except or intersect)
    b) we proceed to spill to disk

However, when we detect this, we have already saved a (too) large blob
row to the hash table (there are two rows in it). The way the overflow
detection works (cf. MEM_ROOT bug fixed in patch two of this series),
is that we detect the overflow once we try to insert the next (third
in this case) row into the hash table after we have exhausted the
available space.  Spill to disk then proceeds to spread the two rows
in the hash table over the chunk files (they end up in different chunk
files).

Then, as we re-load the chunk with the too large row, we fail to load
that chunk with the too large row, although we could save it in the
hash table the first time. The actual reason for this is also
addressed in patch two; we have less space available in the mem_root
than initially due to use of MEM_ROOT::Clear_for_reuse instead of
MEM_ROOT::Clear). In this case, the single too large blob in a row
(713091) far exceeds the (default) size of set_operations_buffer_size
(262144) and triggers overflow on re-loading of the HF chunk file,
which is not supposed to happen, more below.  We then try to move to a
a secondary overflow, which finally asserts because we are in an
unexpected phase: SS_READING_LEFT_HF.  The underlying assumption which
is broken is that once we have spread the rows into the chunk files
(16 in this case), reloading any one of them should always succeed,
and the recovery logic builds on this presumption.

So, in this case, we should not have started spill handling, since it
has no chance of succeeding, in deed it is bound to fail: we cannot
fit a single row into the hash table given the space constraints.

Solution: detect that we have a outsized row as it is initially
entered into the hash table, skip spill handling and move directly to
indexed temporary table de-duplication. For increased safety, we limit
the maximum size of a single row to that of half the
set_operations_buffer_size, cf. the Clear_for_reuse loss mentioned
above.

Change-Id: Id80ba0174d0d587fca68d15a731cd3d60a80b645