Bug#35362984: Memory leak with index merge union plan
Some queries in the federated.federated test leaked memory and caused
failures with ASan. These queries used Rowid-Ordered Retrival index
intersection or union on a table in the Federated storage engine.

The leak was caused by missing cleanup in ha_federated::close(). It
freed only one of the elements in the "results" array before it
cleared the array. Any remaining elements would live on forever as
unreachable objects on the heap.

Fixed by making ha_federated::close() free all the elements of the
array before clearing it. The workarounds for the bug are removed from
the test.

Change-Id: I9a7bd762f2d04e9ff737be0c7069a524cc3e9c4b