Bug#36767881: Assert in Prepared_statement::execute_loop
The problem occurs when there is a prepared statement that calls
a stored procedure. That procedure returns the error ER_NEED_REPREPARE
to Prepared_statement::execute_loop(). Now, the code expects
a reprepare observer to be in invalidated state when this error is
reported, but procedure execution may return this without an
invalidated observer.

The fix is to remove the assert since it is not valid.
The following test is also extended to handle reprepare observers that
are not invalidated, in addition to observers that cannot be retried.

No test cases added, bug was reported from regular test suite with
--ps-protocol parameter.

Change-Id: I476d9fd7716966380123ba1fa51d59102291116d