Bug#35327006: Stabilize the weekly runs [noclose]
Some replication tests fail when running mtr with --hypergraph.

These tests check how replication behaves in certain situations when
statements have to wait for locks, and they are therefore sensitive to
changes in query plans. A change in the query plan may affect which
looks are acquired, and in which order they are acquired, and change
which statements time out or get into a deadlock.

In these tests, a DELETE statement is expected to perform an index
range scan. However, the hypergraph optimizer chooses a table scan
instead. It is a reasonable choice, since the table is very small and
it doesn't really matter which access method is chosen from a cost
perspective. But it changes the order in which the rows are seen and
locked, which is enough to make the test fail.

Fixed by adding an index hint to make the statements use the desired
access method, regardless of which optimizer is used.

Change-Id: Ib95500c344fc0285690631b90b66935a12ec6d1e