Bug#36758378: Signal 11 in Table_ref::is_external
This solution is for MySQL 9.0 and later.

The problem may occur when Sql_cmd_dml::execute() is called, and
we have certain error situations, and we check whether the query
involves an external table. The function has_external_table()
traverses the tables of the query and checks each whether it is
"external". But this involves inspecting the TABLE object, and we
may hit a null pointer since the TABLE pointer inside the Table_ref
object is set to NULL when the table is closed.

This is less vulnerable in 9.0 than in 8.0, since there are fewer
calls to has_external_table(), however this patch makes checking
for external tables foolproof: Instead of traversing all tables whenever
we need to know the value of this property, the value of the property
is stored in the LEX object while opening the tables. We know that the
information is valid at that point in time, and we know the information
is valid as long as the query (and the LEX object) exists.

The test case is from bug#36725524 - although this never failed in 9.0.

Change-Id: I122e75f18d09662e2b63d40982e3e8a624a01489