Bug#35102083: void Query_expression::set_prepared(): Assertion `!is_prepared()' failed.
The problem here may happen when we have a statement with a certain
aggregate function inside a stored routine, and the destination for
the result of the aggregate function is a routine local variable.
In this case, when setting up a target that is usually a temporary
table, instead the optimizer will set up an object with class inherited
from Item_sum_hybrid_field as the target field. These objects are
constructed by the optimizer and evaluated at execution time. When
evaluated, they are also determined to be not resolved, and thus
fix_fields() is called, inside function sp_prepare_func_item().

When this function calls fix_fields(), it also sets the enclosing
query expression (thd->lex->unit) to be "prepared". However, the
expression has no dedicated query expression as it was created as
a specific item during optimization, and thus it attempts to set
the parent query expression as "prepared" which is already the case.
This causes an assert to fire.

It is thus apparent that this mechanism is not designed to be used
with standalone expression objects created during optimization.
But we can make the procedure work: The objects are actually fully
resolved during construction and do not need a call to fix_fields()
call. All they need is setting the "fixed" flag at end of construction.
This means we never reach the part of code where (*it_addr)->fixed
is false, and thus never set the state of the query expression.

Thus, the fix is actually simple. In addition to the bug fix, it was
noted that resolving can be simplified by calling
Item::set_data_type_from_item() instead of setting type members one by
one, which is vulnerable. We also performed a general cleanup by
adjusting class names and member variable names.

Here is a summary of the refactoring performed:

Changed class names:
  Item_sum_hybrid_field -> Item_aggregate_field
  Item_sum_num_field -> Item_aggr_numeric_field
  Item_avg_field -> Item_aggr_avg_field
  Item_sum_bit_field -> Item_aggr_bit_field
  Item_variance_field -> Item_aggr_variance_field
  Item_std_field -> Item_aggr_std_field

Changed member names, class Item_aggregate_field:
  field -> m_field
  hybrid_type -> m_result_type

Changed member names, class Item_aggr_avg_field:
  f_precision -> m_precision
  f_scale -> m_scale
  dec_bin_size -> m_dec_bin_size
  prec_increment -> m_prec_increment

Changed member names, class Item_aggr_bit_field:
  reset_bits -> m_reset_bits

Changed member names, class Item_aggr_variance_field:
  sample -> m_sample

Other changes:
- The resolve_type() function is unused and thus the dummy
  implementation is moved to class Item_aggregate_field.
- Item_sum_hybrid_field::cleanup() is removed since it does no
  reasonable cleanup (the field member is set during resolving and hence
  should not be cleared after a single execution).

Change-Id: I16f7d2a1de9ecdfb53bea96cdeabf8084a652bfe