Bug#35799038 difference in strict mode lead importing tablespace crash
Description:
------------
  The sql_mode determines the syntax checks and data validation
performed by the server. A sql_mode containing 'STRICT_TRANS_TABLE' is
a strict mode. The strict mode influences the calculation of the number
of nullable columns in a secondary index in a table. Hence, a schema
mismatch is possible if the table is created under different sql_mode
even if they have the same definition.
  Importing a tablespace from the IBD file checks for schema mismatch
when a CFG file is provided. However, it also allows the user to skip
the schema checks - this allows one to import the data even if complete
information about the table is unknown.

Issue:
------
  After importing the tablespace, the clustered index is checked for
corruption. However, the secondary indexes are not checked. Just before
returning the success, the statistics of the indexes are updated. In
case there was a schema mismatch, the secondary indexes may have become
corrupt. When updating the statistics, the corrupt indexes are ignored.
However, the indexes were not marked corrupt - this lead to an invalid
page access when trying to refresh statistics of the secondary index.

Fix:
----
  Ensure that other index trees are also checked for corruption after
the tablespace is imported. In case a CFG is provided during the Import,
then reject the import if schema mismatch is observed - even if
nullability of secondary index is different in the CFG file and the
destination table.

Change-Id: Ibdc7ff9bde7028de648a024f403e4f3bcec51a1c