Bug#36790906: MySQL Last_query_cost status incorrect calculated with subqueries [1/2, trace, noclose]
Some queries have been reported to return a very low cost in the
Last_query_cost status variable, even though they are relatively
expensive to execute. This could result in the query being executed in
the primary engine because Last_query_cost is lower than
secondary_engine_cost_threshold, even though the actual cost is well
above the threshold.

This patch adds more information to the optimizer trace about how the
Last_query_cost status variable was calculated, to help diagnose the
problem if it happens again. It is only added to the trace in the
primary-tentatively optimization phase of queries eligible for offload
to a secondary engine, since that's the only case where the
calculation affects how the query is handled.

Change-Id: Ib1794427080e578642ff2a9cf74c7bb53a12c14e