Bug#36784284 - Secondary fails when executing a CREATE TABLE SELECT from an async channel
Running a "CREATE TABLE ... AS SELECT" statement on a source coming
from an asynchronous channel to Group Replication, causes errors
on the replica.

Group replication uses LEX::sql_command to identify the statement
type and SELECT query block checks to identify statement as
"CREATE TABLE ... AS SELECT".

When using "CREATE TABLE ... AS SELECT" for a table with an engine
supporting Atomic DDL, the binary log on the source server will
contain the statement in the following format:

  CREATE TABLE ... START TRANSACTION
  TABLE MAP
  ROW EVENT
  ROW EVENT
  ROW EVENT
  XID / COMMIT

While committing this event (the 'CREATE TABLE ... AS SELECT' statement)
on the group replication node using the XID / COMMIT
(sql_command == SQL_END)statement, logic to identify "CREATE TABLE
... AS SELECT" fails now. This occurs because the conditions for
recognizing the statement as such are not satisfied on the replica.
As consequence, Group Replication includes inaccurate information
in the metadata associated with the event. Hence, applying event
on the secondary in the group replication fails.

To fix this issue, logic to identify "CREATE TABLE ... SELECT"
statement in Group Replication is extended to check if statement
is a transactional DDL. Flag to indicate transactional DDL is set
only while executing "CREATE TABLE ... AS SELECT" statement at
replica.

Change-Id: I19703d8d9040843e45afb031cbc6f86537741618